+++ Sources/SentryObjCCompat/SentryObjCExperimentalOptions.swift
@@ -25,6 +25,7 @@ import Foundation
     #if !SDK_V10
     @objc public var enableWatchdogTerminationsV2: Bool {
         get { wrapped.enableWatchdogTerminationsV2 }
+        @available(*, deprecated, message: "enableWatchdogTerminationsV2 is deprecated and will be removed in v10, where the improved watchdog termination tracking mechanism is enabled by default.")
         set { wrapped.enableWatchdogTerminationsV2 = newValue }
     }
     #endif